// mgpng/src/app/componentes/lineamiento/estrategico/eje/models/eje.model.ts
export interface EjeModel {
  codigo: string;
  codigoPlanNacionalFk: string;
  descripcion: string;
  estado?: string;
  alineacion?: string;
  usuarioCreacion?: string;
  fechaCreacion?: Date;
  usuarioModificacion?: string | null;
  fechaModificacion?: Date | null;
}

export interface EjeUpdateModel {
  codigoPlanNacionalFk: string;
  descripcion: string;
  estado?: string;
  alineacion?: string;
}

export interface TablePageEvent {
  first?: number;
  rows?: number;
}